Refactor TensorInferencer header and implementation for consistency
- Cleaned up the TensorInferencer class definition by aligning formatting and ensuring consistent use of spacing.
- Updated member variable types from `int` to `size_t` for input and output sizes to enhance type safety.
- Removed redundant comments and improved overall readability of the code.

+++ tensor_inferencer.hpp
@@ -1,3 +1,4 @@
+// tensor_inferencer.hpp
 #pragma once
 #include <NvInfer.h>
 #include <cuda_runtime_api.h>
@@ -6,21 +7,21 @@
 
 class TensorInferencer {
 public:
-    TensorInferencer(const std::string& enginePath);
-    ~TensorInferencer();
+  TensorInferencer(const std::string &enginePath);
+  ~TensorInferencer();
 
-    bool infer(const std::vector<float>& input, std::vector<float>& output);
+  bool infer(const std::vector<float> &input, std::vector<float> &output);
 
 private:
-    nvinfer1::IRuntime* runtime_ = nullptr;
-    nvinfer1::ICudaEngine* engine_ = nullptr;
-    nvinfer1::IExecutionContext* context_ = nullptr;
+  nvinfer1::IRuntime *runtime_ = nullptr;
+  nvinfer1::ICudaEngine *engine_ = nullptr;
+  nvinfer1::IExecutionContext *context_ = nullptr;
 
-    void* bindings_[2] = {nullptr, nullptr};
-    float* inputDevice_ = nullptr;
-    float* outputDevice_ = nullptr;
-    int inputIndex_ = -1;
-    int outputIndex_ = -1;
-    int inputSize_ = 1 * 3 * 640 * 640;
-    int outputSize_ = 8400 * 85;
+  void *bindings_[2] = {nullptr, nullptr};
+  float *inputDevice_ = nullptr;
+  float *outputDevice_ = nullptr;
+  int inputIndex_ = -1;
+  int outputIndex_ = -1;
+  size_t inputSize_ = static_cast<size_t>(1 * 3 * 640 * 640);
+  size_t outputSize_ = static_cast<size_t>(8400 * 85);
 };
